+++ Includes~/UnityXRNativePtrs.h
@@ -25,12 +25,12 @@ typedef struct UnityXRNativePlane_1
     void* planePtr;
 } UnityXRNativePlane_1;
 
-// XRReferencePointExtensions.GetNativePtr
-typedef struct UnityXRNativeReferencePoint_1
+// XRAnchor.nativePtr
+typedef struct UnityXRNativeAnchor_1
 {
     int version;
-    void* referencePointPtr;
-} UnityXRNativeReferencePoint_1;
+    void* anchorPtr;
+} UnityXRNativeAnchor_1;
 
 typedef struct UnityXRNativePointCloud_1
 {
@@ -60,7 +60,7 @@ typedef struct UnityXRNativeRaycast_1
 static const int kUnityXRNativeSessionVersion = 1;
 static const int kUnityXRNativeFrameVersion = 1;
 static const int kUnityXRNativePlaneVersion = 1;
-static const int kUnityXRNativeReferencePointVersion = 1;
+static const int kUnityXRNativeAnchorVersion = 1;
 static const int kUnityXRNativePointCloudVersion = 1;
 static const int kUnityXRNativeImageVersion = 1;
 static const int kUnityXRNativeEnvironmentProbeVersion = 1;
@@ -70,7 +70,7 @@ static const int kUnityXRNativeRaycastVersion = 1;
 typedef UnityXRNativeSession_1 UnityXRNativeSession;
 typedef UnityXRNativeFrame_1 UnityXRNativeFrame;
 typedef UnityXRNativePlane_1 UnityXRNativePlane;
-typedef UnityXRNativeReferencePoint_1 UnityXRNativeReferencePoint;
+typedef UnityXRNativeAnchor_1 UnityXRNativeAnchor;
 typedef UnityXRNativePointCloud_1 UnityXRNativePointCloud;
 typedef UnityXRNativeImage_1 UnityXRNativeImage;
 typedef UnityXRNativeEnvironmentProbe_1 UnityXRNativeEnvironmentProbe;