+++ Runtime/ARCoreApi.cs
@@ -7,10 +7,10 @@ namespace UnityEngine.XR.ARCore
     static class Api
     {
 #if UNITY_ANDROID && !UNITY_EDITOR
-        [DllImport("UnityARCore", EntryPoint="UnityARCore_session_setFeatureRequested")]
+        [DllImport(Constants.k_LibraryName, EntryPoint="UnityARCore_session_setFeatureRequested")]
         public static extern void SetFeatureRequested(Feature feature, bool requested);
 
-        [DllImport("UnityARCore", EntryPoint="UnityARCore_session_getRequestedFeatures")]
+        [DllImport(Constants.k_LibraryName, EntryPoint="UnityARCore_session_getRequestedFeatures")]
         public static extern Feature GetRequestedFeatures();
 
         public static bool platformAndroid => true;