+++ Runtime/ArPrestoApi.cs
@@ -5,7 +5,7 @@ namespace UnityEngine.XR.ARCore
     internal static class ArPrestoApi
     {
 #if UNITY_ANDROID && !UNITY_EDITOR
-        [DllImport("UnityARCore")]
+        [DllImport(Constants.k_LibraryName)]
         public static extern void ArPresto_update();
 #else
         public static void ArPresto_update() {}